refactor: Sprint 48 PR 4 — split render.rs into sub-modules
render.rs (2352 LOC) → 7 sub-files in src/render/:
- core_render.rs (672 LOC): render(), tab_bar, status_bar, pane_tree + 8 tests
- border.rs (264 LOC): border grid system + 5 tests
- overlay.rs (428 LOC): menu/rename/help/palette renderers + 5 tests
- panels.rs (565 LOC): decisions/tasks board + 4 tests
- panels_fleet.rs (251 LOC): fleet_view/monitor_view + 1 test
- scratch.rs (62 LOC): scratch shell
- mod.rs (16 LOC): re-exports only

All sub-files ≤672 LOC. Tests distributed inline per sub-module.
Zero behavior change. 1419 tests = baseline.
